Refactor for modularity.
